fix keep alive jobs when no pong response was received yet
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@2245 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/RTMPConnection.java
@@ -969,7 +969,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 				return;
 			}
 
-			if (lastPingSent - lastPongReceived > maxInactivity) {
+			if (lastPongReceived > 0 && lastPingSent - lastPongReceived > maxInactivity) {
 				// Client didn't send response to ping command for too long, disconnect
 				service.removeScheduledJob(keepAliveJobName);
 				keepAliveJobName = null;
